requirements.txt
APScheduler==3.10.4
asttokens==2.4.0
async-timeout==4.0.3
backcall==0.2.0
blinker==1.6.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
Flask==3.0.0
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.2
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
redis==5.0.1
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.10.1
tzdata==2023.3
tzlocal==5.0.1
watchdog==3.0.0
wcwidth==0.2.8
Werkzeug==3.0.0
